Encoder_driver working !!!!
(PID do  not work properly :( )

+++ PinPongGame/Node2Arduino/Encoder_driver.h
@@ -7,11 +7,11 @@
 
 
 
-int32_t PI_controller(int32_t ref, uint32_t pos);
+float PI_controller(int32_t ref, uint32_t pos);
 
 int32_t Motor_position(uint8_t joystick_position, int32_t ref);
 
-void Motor_driving(int32_t u);
+void Motor_driving(float u);
 
 void PWM_Motor_Init();
